Remove Cell Controllers
This commit gets rid of all the cell controllers that I naively created for table view cells. The cell views are not populated with data by tag. I also added an NSXMLParser to the project but I can't seem to get it working correctly - it does not parse all the xml tags in an xml document.