Implemented the Hpple Library
These commit implements the Hpple Library for parsing the HTML from Cinema Center's website. The only issue I've come across so far is that I need to find a better way to concatenate the strings when parsing the movie description. Newline characters are sometimes placed in the wrong spots when I append them myself.